fix: Inject isLoggedIn state from the server
